Simulate right whale movements

Simulate right whale movements

NARW_simulator(
  cohortID,
  seus,
  gsl,
  support,
  densities,
  densities_seus,
  densities_gsl,
  densitySeq,
  latentDensitySeq,
  prey,
  fishing,
  vessels,
  noise,
  doseresp_seed,
  doseresp,
  daylight,
  regions,
  limits,
  limits_daylight,
  limits_regions,
  limits_prey,
  limits_fishing,
  limits_vessels,
  limits_noise,
  resolution,
  resolution_daylight,
  resolution_regions,
  resolution_prey,
  resolution_fishing,
  resolution_vessels,
  resolution_noise,
  M,
  stepsize,
  xinit,
  yinit,
  stressors,
  growth,
  prey_scale,
  starvation,
  starvation_onset,
  nursing_cessation,
  piling_hrs,
  progress
)

Arguments

densities

array of density maps, stored in column-major order with indices [xind, yind, map] for projected spaces (xind is index for a projected lon coordinate, yind is index for a projected lat)

densitySeq

1-indexed vector specifying which density maps should be used to simulate movement at each timepoint; number of timepoints to simulate comes from length of vector densitySeq

latentDensitySeq

1-indexed vector specifying which density maps should be used to simulate movement for each latent animal

limits

vector (xmin, xmax, ymin, ymax) of spatial coordinate extents for spatial densities

resolution

vector (xres, yres) of spatial step sizes for spatial densities

M

Number of proposals used in the importance sampler for movement (Michelot, 2019)

stepsize

Rradius of the proposal circle for movement (Michelot, 2019)

xinit

matrix of initial x coordinates for latent animals (nlatent x n)

yinit

matrix of initial y coordinates for latent animals (nlatent x n)

Value

List of coordinates