Vessel strikes

Generates monthly rasters describing the risk of vessel strikes from wind farm development activities, as defined in a scenario object.

map_vessels(
  obj = NULL,
  z = "dist",
  strike_scalar = 1e-07,
  which.month = 1:12,
  vessel.speed = NULL,
  speed.limit = NULL,
  baseline = TRUE,
  gantt = FALSE,
  do.plot = FALSE,
  spgdf = FALSE
)

Arguments

obj

An object of class narwscenario.

z

Character. Output variable; can be one of "risk" (relative strike risk), "density" (transit density), "transits" (total number of transits per cell), or "dist" (total distance traveled per cell).

strike_scalar

Numeric. Value of the factor used to scale the resulting strike surface(s).

which.month

Integer, or vector of integers. Months of the year for which strike rasters are returned; 1 corresponds to January, 2 to February etc.

vessel.speed

Numeric vector. Nominal transit speeds for each vessel class (Cable Lay, Construction/Crane, Crew Transfer,Heavy Cargo, Support Vessels <100 m, Survey, Tugs).

speed.limit

Numeric vector of length 12, indicating the maximum speeds imposed during slow-downs for each month of the year. Months with NA values are subject to no speed restrictions.

baseline

Logical. Defaults to FALSE. If TRUE, baseline conditions (excluding wind farm development activities) are included in the output.

gantt

Logical. Defaults to FALSE. If TRUE, a Gantt chart showing the schedule of piling is produced.

do.plot

Logical. Defaults to FALSE. If TRUE, strike maps are generated.

spgdf

Defaults to FALSE. If TRUE, output surfaces are returned as SpatialGridDataFrame objects.

Value

A list containing strike risk rasters for each month, in SpatialGridDataFrame or raster format.

Author

Phil J. Bouchet

Examples

if (FALSE) {
library(narwind)
vesselmaps <- map_vessels(obj = scenario_01,
                          z = "dist",
                          strike_scalar = 1e-7,
                          which.month = 1:12,
                          vessel.speed = NULL,
                          speed.limit = c(10,10,10,10,rep(NA,8)),
                          baseline = TRUE,
                          gantt = TRUE,
                          do.plot = TRUE,
                          spgdf = FALSE)
}
...............................
#> Error in eval(expr, envir, enclos): object '...............................' not found
...............................
#> Error in eval(expr, envir, enclos): object '...............................' not found
...............................
#> Error in eval(expr, envir, enclos): object '...............................' not found
...............................
#> Error in eval(expr, envir, enclos): object '...............................' not found
...............................
#> Error in eval(expr, envir, enclos): object '...............................' not found
...............................
#> Error in eval(expr, envir, enclos): object '...............................' not found