Plot population trends

Takes one or more narwproj object(s) and creates plots of the estimated population trajectories.

# S3 method for narwproj
plot(
  ...,
  interval = TRUE,
  cohort = FALSE,
  noaa = FALSE,
  scales = "free",
  ncol = 3,
  nx = 5,
  ny = 5
)

Arguments

...

One or more objects of class narwproj, as returned by predict.narwsim.

interval

Logical. If TRUE, percentile confidence intervals are shown on the plots. Defaults to TRUE.

cohort

Logical. If TRUE, separate plots are returned for each cohort. If FALSE, a single plot for the whole population is shown. Defaults to FALSE.

noaa

Logical. If TRUE, the population trajectory predicted as part of NOAA's population viability analysis (Runge et al., 2023) is also plotted. Defaults to FALSE.

scales

Character. Defines whether axis scales should be constant or vary across plots. Can be one of "fixed" or "free" (the default).

ncol

Integer. Number of columns for the plot layout when cohort = TRUE. Defaults to 3.

nx

Integer. Desired number of x-axis intervals. Non-integer values are rounded down. Defaults to 5. See pretty for details.

ny

Integer. Desired number of y-axis intervals. Non-integer values are rounded down. Defaults to 5. See pretty for details.

References

Runge MC, Linden DW, Hostetler JA, Borggaard DL, Garrison LP, Knowlton AR, Lesage V, Williams R, Pace III RM (2023). A management-focused population viability analysis for North Atlantic right whales. US Dept Commer Northeast Fish Sci Cent Tech Memo 307, 93 p. Available at https://www.fisheries.noaa.gov/s3/2023-10/TM307-508-1-.pdf.

Author

Phil J. Bouchet

Examples

if (FALSE) {
library(narwind)
m <- narw(1000)
m <- augment(m)
p <- predict(m)
plot(p)
}