Model summary and diagnostics

Prints a text-based summary of outputs from the individual-based model to the R console.

# S3 method for narwsim
summary(obj, what = "all", relative = FALSE, plot = FALSE, ...)

Arguments

obj

An object of class narwsim, as returned by narw.

what

Character, or character vector, defining which data to summarize. The default ("all") returns a full summary of all model components. See Details for further information.

relative

Logical. If TRUE, percentages are calculated relative to class totals. Defaults to FALSE.

plot

Logical. If TRUE, additional plots are produced.

...

Optional arguments for subsetting.

Details

An overview of individual model components can be obtained as follows:

what = "health"

Mortality rates by cohort, source, and region

Abortion rates (pregnant females)

Number of calving events and mean/range of birthing dates

what = "movements"

Migratory destinations, by cohort

Step lengths (i.e., daily travel distances) and migration distances

what = "habitat"

Proportion of daily locations by cohort, region, and country

Numbers of animals visiting each region, by cohort

Days spent in each region, by cohort

Total number of regions visited by each individual, by cohort

what = "behavior"

Activity budgets (i.e., time spent traveling vs. feeding vs. resting/nursing), by region and cohort

what = "strike"

Overall strike rates

Strike rates by cohort

Strike probabilities by cohort

what = "gear"

Overall entanglement rates

Entanglement rates by cohort

Attachment sites of gear (body vs. head)

Entanglement frequencies by cohort

Entanglement probabilities by cohort

Entanglement severity by cohort

what = "noise"

Received noise levels, by cohort

Behavioral response thresholds, by cohort

Response durations, by cohort

what = "other"

Overall mortality rates from other sources

Mortality rates from other sources, by cohort

what = "stressors" All outputs from "gear", "strike", "noise", and "other"

what = "energy"

Daily energy intake vs. expenditure, by cohort

Daily energy deficit vs. surplus, by cohort

Author

Phil J. Bouchet

Examples